GraphqlTemplate/requirements.txt
Flask==1.1.4
MarkupSafe==2.0.1
py-healthcheck==1.10.1
alembic==1.8.0
gunicorn==20.1.0
gevent==21.12.0
mypy==0.942
graphene==2.1.9
graphene-stubs==0.15
Flask-GraphQL==2.0.1
requests==2.28.1
pydantic==1.9.1
graphene-federation==0.1.0
python-graphql-client==0.4.3
paho-mqtt==1.6.1
pika==1.3.0
PikaBus==1.3.6
graphql-ws==0.4.4
flask-sockets==0.2.1